extract to ModuleLaunchPad component, add test
